Fixed bug where child was always re-rendered
Whenever a parent component was changed, it was re-rendering and diff all of the child components in the tree regardless of whether or not they had changed.

This was because it was both force updating the child components in the diff and the shouldUpdate method was wrong.

Now when a parent changes it will check the child components properly and only update the sub-trees it needs to.