No longer flattening children in virtual nodes
This was to get around a user error when not using JSX. People would do something like `dom('div', null, [props.children])`  and not realize the `props.children` is an array.

In ES6 we can just do `dom('div', null, [...props.children])```.

Using an array as a child node now throws an error.