Nested components don't share an element
This was requiring some hacky work to get components down the tree who have components as their root, to all have the same DOM element reference in afterMount. This el is mainly used to get around DOM side-effects, like focusing and element or doing something with the rendered dimensions of the DOM element.

We can find a better way to do this that doesn't couple DOM element lifecycle events with component lifecycle events.

DOM element attributes can now be a function that will be called with the element. Attributes that map to events will work as normal, but attributes with functions will be called every time the attribute value changes.