Entities are now a tree
The component tree is now a real tree, instead of having an entities hash and a children hash. We originally had this because we thought it would be nice to loop through all of the entities, or be able to reference an entity in a different part of the tree. This allowed up to walk up the tree, but we can always store the owners in another hash later if we need to because the owner can never change.